monitoring_suite: Split large one file test into each resource test
