+++ internal/service/monitoring_suite/data_source_trace_storage_test.go
@@ -1,0 +1,47 @@
+// Copyright 2016-2026 The terraform-provider-sakura Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package monitoring_suite_test
+
+import (
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+	"github.com/sacloud/terraform-provider-sakura/internal/test"
+)
+
+func TestAccSakuraMonitoringSuiteTraceStorageDataSource_basic(t *testing.T) {
+	resourceName := "data.sakura_monitoring_suite_trace_storage.foobar"
+	rand := test.RandomName()
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { test.AccPreCheck(t) },
+		ProtoV6ProviderFactories: test.AccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			{
+				Config: test.BuildConfigWithArgs(testAccSakuraMonitoringSuiteTraceStorageDataSource_basic, rand),
+				Check: resource.ComposeTestCheckFunc(
+					test.CheckSakuraDataSourceExists(resourceName),
+					resource.TestCheckResourceAttr(resourceName, "name", rand),
+					resource.TestCheckResourceAttr(resourceName, "description", "description"),
+					resource.TestCheckResourceAttrSet(resourceName, "created_at"),
+					resource.TestCheckResourceAttrSet(resourceName, "account_id"),
+					resource.TestCheckResourceAttrSet(resourceName, "resource_id"),
+					resource.TestCheckResourceAttrSet(resourceName, "retention_period_days"),
+					resource.TestCheckResourceAttrSet(resourceName, "endpoints.ingester.address"),
+				),
+			},
+		},
+	})
+}
+
+var testAccSakuraMonitoringSuiteTraceStorageDataSource_basic = `
+resource "sakura_monitoring_suite_trace_storage" "foobar" {
+  name = "{{ .arg0 }}"
+  description = "description"
+}
+
+data "sakura_monitoring_suite_trace_storage" "foobar" {
+  id = sakura_monitoring_suite_trace_storage.foobar.id
+}
+`
